mega_code/client/pyproject.toml
[project]
name = "mega-code"
version = "1.0.0"
description = "Open-source Claude Code plugin — collect interaction data, extract skills, and optimize AI workflows"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = "MIT"
authors = [
    { name = "Mind AI", email = "support@megacode.ai" }
]
keywords = ["claude-code", "plugin", "skills", "optimization", "data-collection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
parquet = ["pyarrow>=12.0"]
telemetry = ["opentelemetry-api>=1.20.0", "opentelemetry-sdk>=1.20.0"]

[project.urls]
Homepage = "https://megacode.ai"
Repository = "https://github.com/wisdomgraph/mega-code"

[project.scripts]
mega-code = "mega_code.client.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mega_code"]
only-include = [
    "mega_code/__init__.py",
    "mega_code/client",
]
artifacts = [
    "*.toml",
    ".claude-plugin/**",
    "hooks/**",
    "commands/**",
    "skills/**",
]

[tool.hatch.build.targets.wheel.force-include]
"pyproject.toml" = "mega_code/data/pyproject.toml"
"README.md" = "mega_code/data/README.md"
".claude-plugin" = "mega_code/data/.claude-plugin"
"hooks" = "mega_code/data/hooks"
"commands" = "mega_code/data/commands"
"skills" = "mega_code/data/skills"

[tool.hatch.build.targets.sdist]
include = [
    "mega_code/__init__.py",
    "mega_code/client/",
    ".claude-plugin/",
    "hooks/",
    "commands/",
    "skills/",
    "README.md",
    "pyproject.toml",
]
